Reworked logo to have only 22% of previous size
Instead of bunch of ‘pixel’ rectangles I've defined a pattern and created the ‘pixel-heart’ path which is filled with the pattern.

The image is rescaled to match 100 pixels size for pixel perfection (browsers has troubles with rasterizing and scaling patterns). There are still some blurriness in certain conditions, but I think it has Ok quality now.

There is half a pixel shift of the path to prevent thin red spans due to rounding errors which can occur in some non-100% scale.